timing: make nsPerTick precise, stop truncating it to int
comp->nsPerTick (real ns per emulated T-state) was truncated to int,
which compounded through nsPerDot/nsPerLine/nsPerFrame and made every
machine family run 0.3-0.9% faster than real hardware (audio pitch,
frame rate, everything paced off it). Made it double throughout, and
round once from the precise value instead of compounding an
already-truncated int; two ns/tick divisions get llround() since
division can flip a whole tick due to float rounding where
multiplication can't. Dropped the now-unneeded even-masking of
nsPerTick in zx_init.

Also bumped the CPU-frequency spinbox in Options from 2 to 4 decimals:
Qt rounds the stored value to `decimals`, so opening Options and
clicking OK was silently collapsing e.g. 3.5469 to 3.55.

+++ src/libxpeccy/spectrum.h
@@ -129,7 +129,9 @@ typedef struct Computer {
 	int frmtCount;		// accumulate T, but reset each INT
 	int hCount;		// T before HALT = frmtCount @ HALT
 	int fCount;		// T in last frame
-	int nsPerTick;
+	double nsPerTick;	// real ns/T, kept precise: truncating this to int was the
+				// root cause of a systemic ~0.3-0.9% emulation speed error
+				// (see nsPerSample-driven pacing in ethread.cpp)
 
 	bool flag[128];			// each machine have its own flags
 	bool sysflag[32];		// some common flags used by several machines or debuga
